+++ ios/AppstackBridge.h
@@ -26,6 +26,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 + (NSString *)getAppstackId;
 
++ (BOOL)isSdkDisabled;
+
 @end
 
 NS_ASSUME_NONNULL_END