+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the React Native Appstack SDK will be documented in this 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.0.22] - 2025-11-06
+
+- **Add a isSdkDisabled() method to know if the SDK is disabled in one of the bridges**
+- **Appstack iOS SDK new dependency: 2.6.3**
+- **Appstack Android SDK new dependency: 0.0.15**
+
 ## [0.0.21] - 2025-10-29
 
 - **Patch the getAppstackId() method to retrieve the install ID of a user without errors**